requirements.txt
fastapi
uvicorn
pydantic
requests
qrcode[pil]
fpdf2